+++ src/rics/performance/__init__.py
@@ -2,7 +2,7 @@
 
 from ._format_perf_counter import format_perf_counter, format_seconds
 from ._multi_case_timer import MultiCaseTimer
-from ._util import plot_run, to_dataframe
+from ._util import get_best, plot_run, to_dataframe
 from ._wrapper import run_multivariate_test
 
 __all__ = [
@@ -12,4 +12,5 @@ __all__ = [
     "format_seconds",
     "plot_run",
     "to_dataframe",
+    "get_best",
 ]