+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
+- Duplicate derived names for candidates in for perftests.
 - Improve handling for optional `Format` blocks (#52). Nested optional blocks are *not* supported.
 
 ## [0.15.0] - 2022-08-01